Fix JSON/YAML output bug
As described in #33, when the output format is set to YAML or JSON and
the installed Chart was found in multiple repositories, then `whatup`
wouldn't show these Charts in the JSON/YAML output.

This was because the only the `releases` field of the
`outdatedListWriter` struct was given as argument to the
`EncodeJSON(..)` func.

Fixes #33