Add row/field 'REPOSITORY' to output
The `REPOSITORY` row/field is added to provide the user the possibility
to easily check which repository does contain the newer version of the
installed chart.

It was implemented not only to improve UX but also to prevent bug
reports in the future.
Image if you install a Chart `mysql` from the `server1` repository and
after a while `server1` gets deprecated and now there exists `server2`.
The user now has to check which repository does contain the specific
version by running `helm repo search […]`.