Fix showing deprecated repos in multi-repo view
If multiple repos do serve a chart AND the user defined
`--ignore-deprecation=true` `whatup` would show the latest chart version
of the repository which is NOT marked as "deprecated".

Output:
    NAME                       postgresql
    NAMESPACE                  devops
    INSTALLED CHART VERSION    6.2.1
    INSTALLED APP VERSION      11.10.0

    REPOSITORY	DEPRECATED	CHART VERSION	APP VERSION
    bitnami   	false     	10.2.5       	11.10.0
    stable    	false      	8.6.2        	11.7.0

Example:
    Repository "stable" serves chart "postgresql"

    | Chart Version | Deprecated |
    | ------------- | ---------- |
    |     8.6.1     |   false    |
    |     8.6.2     |   false    |
    |     8.6.3     |   true     |
    |     8.6.4     |   true     |

    Repository "bitnami" serves chart "postgresql"

    | Chart Version | Deprecated |
    | ------------- | ---------- |
    |     10.2.3    |   false    |
    |     10.2.4    |   false    |
    |     10.2.5    |   false    |

The bug is pretty simple: Whatup does show the latest chart version of
the "stable" repository which is NOT marked as "deprecated".
It was caused because we skipped the tracking code in `searchChart(…)`
if the flag was set to `true` and the chart was marked as `deprecated`.
Previous versions, which haven't been marked as "deprecated" are already
tracked.

To solve it, we store the information if it's marked as deprecated in
the `repo` map and check this value if we have multiple repos serving
the repository.
This check and the `goto` (`SJMP`) statement do introduce a minimal
overhead which will have not measurable effect of the execution time.

Ref #36
Ref https://github.com/fabmation-gmbh/helm-whatup/issues/36#issuecomment-768978347